aportes servicios de carlos

+++ src/main/java/co/edu/cesde/ga/repository/service/StudentService.java
@@ -10,7 +10,7 @@ public interface StudentService {
 
     boolean update(Student student);
 
-    boolean delete(Integer id);
+    boolean delete(Long id);
 
     Student findById(Long studentId);
